+++ src/pages/fieldManagement/settings/FieldTypeSelector.jsx
@@ -6,7 +6,6 @@ import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import InputLabel from '@material-ui/core/InputLabel';
 
-import { fieldTypeChoices } from '../../../constants/fieldTypes';
 import fieldTypes, {
   fieldTypeInfo,
 } from '../../../constants/fieldTypesNew';
