+++ src/pages/fieldManagement/settings/saveField/getTypeCategories.js
