+++ src/pages/serverStatus/components/LegendItem.jsx
