+++ src/pages/serverStatus/components/WaffleSquare.jsx
