+++ src/pages/userManagement/UserManagement.jsx
@@ -13,11 +13,9 @@ import Button from '../../components/Button';
 import Text from '../../components/Text';
 import usePostUser from '../../models/users/usePostUser';
 
-export default function AdminActions() {
+export default function UserManagement() {
   const intl = useIntl();
-  useDocumentTitle(
-    intl.formatMessage({ id: 'ADMINISTRATIVE_ACTIONS' }),
-  );
+  useDocumentTitle(intl.formatMessage({ id: 'MANAGE_USERS' }));
 
   const {
     postUser,
@@ -32,14 +30,13 @@ export default function AdminActions() {
   const [newUserPassword, setNewUserPassword] = useState('');
   const [deleteUserEmail, setDeleteUserEmail] = useState('');
   const [deleteUsername, setDeleteUsername] = useState('');
-  const [restoreEncounterId, setRestoreEncounterId] = useState('');
   return (
     <MainColumn>
       <Text
         variant="h3"
         component="h3"
         style={{ padding: '16px 0 16px 16px' }}
-        id="ADMINISTRATIVE_ACTIONS"
+        id="MANAGE_USERS"
       />
       <Grid
         container
@@ -170,41 +167,6 @@ export default function AdminActions() {
                 style={{ backgroundColor: '#ff171b', color: 'white' }}
               >
                 <FormattedMessage id="DELETE_ACCOUNT" />
-              </Button>
-            </div>
-          </Paper>
-        </Grid>
-
-        <Grid item style={{ width: '100%' }}>
-          <Text
-            variant="h6"
-            style={{ marginTop: 20, marginLeft: 12 }}
-            id="RESTORE_DELETED_ENCOUNTER"
-          />
-          <Paper
-            elevation={2}
-            style={{
-              marginTop: 20,
-              marginBottom: 12,
-              padding: 24,
-              display: 'flex',
-              flexDirection: 'column',
-            }}
-          >
-            <Text id="RESTORE_DELETED_ENCOUNTER_INSTRUCTIONS" />
-            <div style={{ display: 'flex' }}>
-              <LabeledInput
-                schema={{
-                  labelId: 'ENCOUNTER_ID',
-                  displayType: 'string',
-                }}
-                value={restoreEncounterId}
-                onChange={setRestoreEncounterId}
-              />
-            </div>
-            <div style={{ marginTop: 8 }}>
-              <Button display="primary">
-                <FormattedMessage id="RESTORE_ENCOUNTER" />
               </Button>
             </div>
           </Paper>
